Replace FlowNodeInstanceList with Array<FlowNodeInstance> in DataModels
