+++ src/apis/iflow_node_instance_management_api.ts
@@ -1,8 +1,8 @@
 import {IIdentity} from '@essential-projects/iam_contracts';
 
-import {FlowNodeInstanceList} from '../data_models/flow_node_instance';
+import {FlowNodeInstance} from '../data_models/flow_node_instance';
 
 export interface IFlowNodeInstanceManagementApi {
 
-  getFlowNodeInstancesForProcessInstance(identity: IIdentity, processInstanceId: string): Promise<FlowNodeInstanceList>;
+  getFlowNodeInstancesForProcessInstance(identity: IIdentity, processInstanceId: string): Promise<Array<FlowNodeInstance>>;
 }
