+++ src/data_models/flow_node_instance/event_type.ts
@@ -1,0 +1,8 @@
+export enum EventType {
+  linkEvent = 'linkEvent',
+  messageEvent = 'messageEvent',
+  signalEvent = 'signalEvent',
+  terminateEvent = 'terminateEvent',
+  timerEvent = 'timerEvent',
+  errorEvent = 'errorEvent',
+}
