--- src/data_models/flow_node_instance/flow_node_instance_list.ts
@@ -1,7 +1,0 @@
-import {FlowNodeInstance} from './flow_node_instance';
-
-export class FlowNodeInstanceList {
-
-  public flowNodeInstances: Array<FlowNodeInstance>;
-
-}
