+++ src/data_models/flow_node_instance/index.ts
@@ -1,1 +1,1 @@
-export * from './flow_node_instance_list';
+export * from './flow_node_instance';
